+++ src/main/java/com/codeup/closetrycorner/services/UsersRepo.java
@@ -3,7 +3,7 @@ package com.codeup.closetrycorner.services;
 import com.codeup.closetrycorner.models.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UsersRepo extends CrudRepository<User, Integer> {
+public interface UsersRepo extends CrudRepository<User, Long> {
 
     User findByUsername(String username);
